% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{apply_relative_extent}
\alias{apply_relative_extent}
\title{Compute a domain extent from a process extent within a domain extent.}
\usage{
apply_relative_extent(domain_extent, process_extent)
}
\arguments{
\item{domain_extent}{The larger bounding box.}

\item{process_extent}{A bounding box within the domain extent.}
}
\value{
A bounding box measured from the domain extent's coordinate system.
}
\description{
Compute a domain extent from a process extent within a domain extent.
}
