% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{ar_of_pr_rho}
\alias{ar_of_pr_rho}
\title{Creates a function that interpolates an AR-to-PR dataset.
The data calculates attack rate as a function of PR and recovery.}
\usage{
ar_of_pr_rho(pr_to_ar_dt)
}
\arguments{
\item{pr_to_ar_dt}{A data.table with columns AR, PR, and rho.}
}
\value{
A function with signature function(pr, rho) -> AR.
}
\description{
Creates a function that interpolates an AR-to-PR dataset.
The data calculates attack rate as a function of PR and recovery.
}
