% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdline.R
\name{arg_parser}
\alias{arg_parser}
\title{Parse command-line arguments.}
\usage{
arg_parser(args = NULL)
}
\arguments{
\item{args}{Optionally pass the results of \code{commandArgs(TRUE)}.
We have this parameter so that we can test without side effects.}
}
\value{
A list where args that aren't on the command line are NULL.
}
\description{
Parse command-line arguments.
}
