% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{assemble}
\alias{assemble}
\title{Reads output from workers and produces GeoTIFFs and images.}
\usage{
assemble(args)
}
\arguments{
\item{args}{Command-line arguments. Ignores the task but uses the number of tasks.

If you want to call this, see \code{\link{main}} for an example.}
}
\description{
Reads output from workers and produces GeoTIFFs and images.
}
