% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{build_ar2pr}
\alias{build_ar2pr}
\title{This goes the other way, from AR to PR, with rho=0.}
\usage{
build_ar2pr(pr_ar_data)
}
\arguments{
\item{pr_ar_data}{This is the table of pr, ar, and rho values.}
}
\description{
This goes the other way, from AR to PR, with rho=0.
}
