% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{check_args}
\alias{check_args}
\title{Check whether the input arguments make sense.}
\usage{
check_args(args)
}
\arguments{
\item{args}{A list of command-line arguments resulting from the arg parser.}
}
\value{
The same list of command-line arguments with some new types.
}
\description{
Check whether the input arguments make sense.
}
