% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_work.R
\name{check_linearized_outputs}
\alias{check_linearized_outputs}
\title{Check that the results are all finite.}
\usage{
check_linearized_outputs(results, inputs)
}
\arguments{
\item{results}{A list of arrays.}

\item{inputs}{The input list of arrays, used for debugging failure.}
}
\value{
The input results. This will stop on failure.
We shouldn't have any NANs here. Good data in, good data out.
This is what let's us ignore NA later. This function is expensive
because it scans all points, but it's important.
}
\description{
Check that the results are all finite.
}
