% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_work.R
\name{chunk_nonzero}
\alias{chunk_nonzero}
\title{Find nonzeros in each chunk of tile data to see they loaded.}
\usage{
chunk_nonzero(chunks)
}
\description{
Find nonzeros in each chunk of tile data to see they loaded.
}
