% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_work.R
\name{chunk_sums}
\alias{chunk_sums}
\title{Find the sum of each chunk of tile data to see they differ.}
\usage{
chunk_sums(chunks)
}
\arguments{
\item{chunks}{A list of tiles with data.}
}
\description{
Find the sum of each chunk of tile data to see they differ.
}
