% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{collect_and_permute}
\alias{collect_and_permute}
\title{Given a list of matrices, construct a single three-dimensional array.}
\usage{
collect_and_permute(list_of_matrices)
}
\arguments{
\item{list_of_matrices}{Each item in the list is a two-dimensional array of the same size.}
}
\value{
A single three-dimensional array where the first dimension is the
number of items in the list and the next two are the dimensions of each array.
}
\description{
Given a list of matrices, construct a single three-dimensional array.
}
