% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{combine_output}
\alias{combine_output}
\title{Given everything from lapply, combine it.}
\usage{
combine_output(tile_output, blocksize, only_name = NULL)
}
\arguments{
\item{tile_output}{is a list where each entry is a all data for a single
tile. Each entry has one member named \code{block}, which identifies the
tile. The ther entries are the arrays in (year, row, col).}

\item{blocksize}{An named array with "row" and "col".}

\item{only_name}{The name of a single array to combine from all tiles.}
}
\value{
The data combined, where each piece is offset by the same offset.
}
\description{
Given everything from lapply, combine it.
}
