% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{construct_plan}
\alias{construct_plan}
\title{Makes an initial plan.}
\usage{
construct_plan(args)
}
\arguments{
\item{args}{Command-line arguments. Doesn't care about the number of tasks,
but otherwise might as well use the same command-line arguments here
as elsewhere.

If you want to call this, see \code{\link{main}} for an example.}
}
\description{
Makes an initial plan.
}
