% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_work.R
\name{draw_parameters}
\alias{draw_parameters}
\title{Takes a parameter that is an expression (of drawing a distribution) and calls it.}
\usage{
draw_parameters(parameters, N)
}
\description{
Takes a parameter that is an expression (of drawing a distribution) and calls it.
}
