% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_work.R
\name{draw_parameters_three}
\alias{draw_parameters_three}
\title{Takes a parameter that is an expression (of drawing a distribution) and calls it.}
\usage{
draw_parameters_three(parameters, N)
}
\arguments{
\item{parameters}{The config file parameters, no draws on input.}

\item{N}{the number of draws to make. Can be 1 for no draws.}
}
\value{
A data frame with one row per draw, even if it's one.
Goes with pixel_three.
}
\description{
Takes a parameter that is an expression (of drawing a distribution) and calls it.
}
