% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_work.R
\name{draw_parameters_two}
\alias{draw_parameters_two}
\title{Takes a parameter that is an expression (of drawing a distribution) and calls it.}
\usage{
draw_parameters_two(parameters, N)
}
\arguments{
\item{parameters}{A list of floats and integers.}

\item{N}{The integer number of parameters to draw. Can be 1 to use means.}
}
\description{
Takes a parameter that is an expression (of drawing a distribution) and calls it.
}
