% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{flatten_draws}
\alias{flatten_draws}
\title{The draws are hierarchical by variable and then quantile. This flattens them.}
\usage{
flatten_draws(draw_list)
}
\arguments{
\item{draw_list}{The hierarchical list with names.}
}
\value{
A single list with the two names concatenated.
}
\description{
The draws are hierarchical by variable and then quantile. This flattens them.
}
