% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_outline.R
\name{gadm_country_shapefile}
\alias{gadm_country_shapefile}
\title{Retrieve the shapefile for a country at the given level.
GADM identifies countries with the Alpha-3 country code.
If the file doesn't exist, it will return a NULL.}
\usage{
gadm_country_shapefile(alpha3, admin_level = 0, gadm_version = "201104")
}
\arguments{
\item{alpha3}{The three-letter country code.}

\item{admin_level}{0 is the whole country. Different countries have more or
fewer levels.}

\item{gadm_version}{When this version of GADM was downloaded.}
}
\value{
A shapefile from the \code{sf} package, read with \code{st_read}.

This function depends on having a working \code{tempfile()} command because it
has to unzip the shapefile into a temporary directory.
}
\description{
Retrieve the shapefile for a country at the given level.
GADM identifies countries with the Alpha-3 country code.
If the file doesn't exist, it will return a NULL.
}
