% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{improved_errors}
\alias{improved_errors}
\title{Asks R for a stack trace that shows which function had the problem.}
\usage{
improved_errors()
}
\description{
https://renkun.me/2020/03/31/a-simple-way-to-show-stack-trace-on-error-in-r/
}
