% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_work.R
\name{linearized_work}
\alias{linearized_work}
\title{Takes an input array in multiple dimensions and linearizes and removes NA.
This replaces the outputs back into multiple dimensions after making the call.}
\usage{
linearized_work(input_list, run_func)
}
\description{
Takes an input array in multiple dimensions and linearizes and removes NA.
This replaces the outputs back into multiple dimensions after making the call.
}
