% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{load_data}
\alias{load_data}
\title{Load input data, including pr2ar, the PfPR, and the AM.}
\usage{
load_data(config, pr2ar_version, domain_extent, years)
}
\arguments{
\item{config}{Input configuration file name.}

\item{pr2ar_version}{Which version of the \code{pr2ar_mesh.csv} file.}

\item{domain_extent}{Bounding box that we compute within the MAP raster.}

\item{years}{Which years to do.}
}
\value{
A list with the loaded data.
}
\description{
Load input data, including pr2ar, the PfPR, and the AM.
}
