% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_compare.R
\name{load_pixel_script}
\alias{load_pixel_script}
\title{Loads a script that has functions to compute pixel values.}
\usage{
load_pixel_script(script_name)
}
\arguments{
\item{script_name}{The file with the script.}
}
\description{
Loads a script that has functions to compute pixel values.
}
