% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_io.R
\name{load_plan}
\alias{load_plan}
\title{Load the plan that you saved.}
\usage{
load_plan(plan_file)
}
\arguments{
\item{plan_file}{The filename containing the plan.}
}
\description{
Load the plan that you saved.
}
\seealso{
\code{\link{save_plan}}
}
