% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{my_tasks}
\alias{my_tasks}
\title{Pull my work, according to my task index.}
\usage{
my_tasks(task_id, task_cnt, work)
}
\arguments{
\item{task_id}{integer index of task within all tasks. One-based.}

\item{task_cnt}{integer total number of tasks.}

\item{work}{Data frame with row and column of tile.}
}
\value{
2D array that is (row, col) by work items.

The math here is so that, if the tasks don't split evenly, there
isn't a single task with only one thing to do. The first few
tasks get one more, and the rest get one less.
}
\description{
Pull my work, according to my task index.
}
