% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{over_block}
\alias{over_block}
\title{Accumulate the individual time series from a computation that doesn't care about time.}
\usage{
over_block(chunk, parameters, pr_to_ar_dt)
}
\arguments{
\item{chunk}{A list with pfpr, am, and the tile.}

\item{parameters}{A list of parameters for the calculation.}

\item{pr_to_ar_dt}{The pr-ar data from the mechanistic model.}
}
\value{
a list with an entry for each of alpha, kappa, eir, vc, rc.
It also identifies the source block by block_id.

If the pixel function doesn't care about time, then there is no
reason to loop over the matrix. Turn the matrix into a single vector
and then do the computation.
}
\description{
Accumulate the individual time series from a computation that doesn't care about time.
}
