% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_work.R
\name{over_block_draw}
\alias{over_block_draw}
\title{Run the pixel function on a chunk of data.}
\usage{
over_block_draw(chunk, parameters, pr_to_ar_dt, confidence_percent)
}
\arguments{
\item{chunk}{a list with the pfpr and am datasets.}

\item{parameters}{a data.frame, at this point, with a row per draw.}

\item{pr_to_ar_dt}{The pr2ar mesh data.}

\item{confidence_percent}{Confidence interval, out of 100.}
}
\description{
Run the pixel function on a chunk of data.
}
