% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_variables.R
\name{pixel_bounding_box}
\alias{pixel_bounding_box}
\title{Given a bounding box in map coordinates (maybe lat-long), return pixel bbox.}
\usage{
pixel_bounding_box(map_raster, bbox)
}
\arguments{
\item{map_raster}{The \code{raster} package raster for the pixels.}

\item{bbox}{A bounding box as returned by \code{sf::st_bbox}.}
}
\value{
An extent, in pixels.
}
\description{
Given a bounding box in map coordinates (maybe lat-long), return pixel bbox.
}
