% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_work.R
\name{pixel_three}
\alias{pixel_three}
\title{This version works harder on the low end and high end of PfPR
to make the values have good variance there.}
\usage{
pixel_three(pfpr, am, params, strategies)
}
\arguments{
\item{pfpr}{A list of pfpr values.}

\item{am}{A list of treatment values, the same length as pfpr.}

\item{params}{either a list or data frame, to be used in a with-statement.}

\item{strategies}{A list of functions to call to do parts of the work.}
}
\value{
a list with a member for each variable.
}
\description{
This version works harder on the low end and high end of PfPR
to make the values have good variance there.
}
