% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_work.R
\name{pixel_two}
\alias{pixel_two}
\title{Central calculation for pixel work that has low Rc.}
\usage{
pixel_two(pfpr, am, params, strategies)
}
\arguments{
\item{pfpr}{parasite rate.}

\item{am}{anti-malarial treatment rate.}

\item{params}{Parameters for the math.}

\item{strategies}{Functions to apply during calculation.}
}
\value{
Vectors of results.
This was made with fits to the middle values of PfPR.
It was finished Monday 23 Nov 2020. This was the version sent
to MAP that had reasonable values (c is corrected in last step).
}
\description{
Central calculation for pixel work that has low Rc.
}
