% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{pixel_work}
\alias{pixel_work}
\title{Compute the results of analyzing a time series of PfPR and AM.}
\usage{
pixel_work(pfpr, am, params, strategies)
}
\arguments{
\item{pfpr}{A numeric vector of PR. Length is number of time points.}

\item{am}{A numeric vector of case management measure. Length is number of time points.}

\item{params}{Other parameters for the calculation in a list.}

\item{pr2ar}{A function that gives AR from PR and rho.}
}
\value{
A list where each variable has results for the time series
as a vector of numeric. The entries are \code{alpha}, \code{kappa}, \code{eir}, \code{vc}, \code{rc}.

The parameters passed into this function are defined in the config
file, whose default name is \code{rc_kappa.toml}, in the \verb{[parameters]} section.
}
\description{
Compute the results of analyzing a time series of PfPR and AM.
}
