% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{plan_domain_decomposition}
\alias{plan_domain_decomposition}
\title{Given command-line arguments, determine how much data to load.
It's deciding splits without consulting arguments for now.
This decomposes the domain into tiles because, when we parallelize
with the parallel package, each task needs at least a few seconds
of work to do in order to make data transfer worthwhile.}
\usage{
plan_domain_decomposition(available, options)
}
\arguments{
\item{available}{A description of what data could be processed.}

\item{options}{These parameters define \code{blocksize} and \code{single_tile_max}.}
}
\value{
The same list with an added \code{tiles} member that describes
the dimensions of computational tiles relative to the origin
of the \code{domain_extent}, the portion we will compute.
}
\description{
Given command-line arguments, determine how much data to load.
It's deciding splits without consulting arguments for now.
This decomposes the domain into tiles because, when we parallelize
with the parallel package, each task needs at least a few seconds
of work to do in order to make data transfer worthwhile.
}
