% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_variables.R
\name{plot_as_png}
\alias{plot_as_png}
\title{Customized plotting for different variables.}
\usage{
plot_as_png(raster_obj, filename, name, year, options, admin0)
}
\arguments{
\item{raster_obj}{A \code{raster::raster} object.}

\item{filename}{A string filename for the png output.}

\item{name}{The variable and year, with an underscore between. Has to be this.}

\item{year}{The numeric year.}

\item{options}{A list of options for png width, height, and resolution.}

\item{admin0}{A shapefile with admin0 boundaries.}
}
\description{
Customized plotting for different variables.
}
