% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{prepare_timeseries}
\alias{prepare_timeseries}
\title{Given the raw input data, prepare it for computation.}
\usage{
prepare_timeseries(data, work, process_extent, blocksize)
}
\arguments{
\item{data}{A list where each item is input data.}

\item{work}{The matrix of which tiles to do.}

\item{process_extent}{Bounding box of loaded data within the computational domain.}

\item{blocksize}{The size of each tile.}
}
\value{
A list with transformed data items. This list is broken
up into the work tiles, so that we don't transfer the whole
image to each processor when we do calculations.

This is data movement to set up for the next algorithm.
}
\description{
Given the raw input data, prepare it for computation.
}
