% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extents.R
\name{raster_extent_from_work}
\alias{raster_extent_from_work}
\title{Given a list of tiles, figures out the bounding box of pixel values.}
\usage{
raster_extent_from_work(work, blocksize)
}
\arguments{
\item{work}{A 2d matrix with a dimname on the first dimension of "row", "col".}

\item{blocksize}{A row and col size for blocks for the work.}
}
\description{
Given a list of tiles, figures out the bounding box of pixel values.
}
