% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{read_dataset_names}
\alias{read_dataset_names}
\title{Finds the name of every three-dimensional array in a file.}
\usage{
read_dataset_names(filename)
}
\arguments{
\item{filename}{The HDF5 file to read.}
}
\description{
Finds the name of every three-dimensional array in a file.
}
