% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{remove_tiles_over_water}
\alias{remove_tiles_over_water}
\title{Removes tiles that have all NA for this example slice.
Tile is within the domain. The sample data is already cut
to that domain.}
\usage{
remove_tiles_over_water(available)
}
\description{
Removes tiles that have all NA for this example slice.
Tile is within the domain. The sample data is already cut
to that domain.
}
