% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serialize.R
\name{save_chunks}
\alias{save_chunks}
\title{Save a list of arrays to a single file.}
\usage{
save_chunks(filename, chunks, group_name = "chunks")
}
\arguments{
\item{filename}{The filename to open, save, and close.}

\item{chunks}{A list, where each member is an array,
and each array has dimnames where each dimension is
named. Any attributes of those arrays will also be
saved, if HDF5 will save them. I don't know the rules
on what it will/won't save, but strings and arrays
should be good.

This function is for saving arrays in HDF5.
We need to save chunks of arrays, with lots of metadata,
so HDF5 lets us do that and keep everything together,
instead of trying to parse filenames to do it.}
}
\description{
Save a list of arrays to a single file.
}
