% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_io.R
\name{save_plan}
\alias{save_plan}
\title{Save a plan for parallel execution to a file.}
\usage{
save_plan(plan, plan_file)
}
\arguments{
\item{plan}{is a list with some members: \code{domain_extent}, \code{domain_dimensions},
and a list \code{list(tiles = list(blocksize))}.}

\item{plan_file}{is a filename.

This uses JSON because the named lists didn't save correctly in
YAML, and TOML requires your having Python installed, so that's out.}
}
\description{
Save a plan for parallel execution to a file.
}
