% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_main.R
\name{slice_main}
\alias{slice_main}
\title{If you want to run from the Bash command line, you can call this.}
\usage{
slice_main()
}
\description{
If you want to run from the Bash command line, you can call this.
}
