% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_work.R
\name{summarize_draws}
\alias{summarize_draws}
\title{Given a list of draws, summarize mean and confidence intervals.}
\usage{
summarize_draws(draws, confidence_percent)
}
\arguments{
\item{draws}{A list of draws. Each draw is a list of variables,
where each variable is a three-dimensional array.}

\item{confidence_percent}{As in, 95.}
}
\value{
a list with one name for each variable. Each list
item has three parts, lower, median, and upper, for that variable.
Each of lower, median, and upper has dimensions year, row, col.
}
\description{
Given a list of draws, summarize mean and confidence intervals.
}
