% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_decomposition.R
\name{task_work}
\alias{task_work}
\title{Takes the extent of the tiles and returns a list of which tiles to do.}
\usage{
task_work(tile_cnt)
}
\arguments{
\item{tile_cnt}{The tile_cnt from the plan from \code{plan_domain_decomposition}.}
}
\value{
An array that's 2 x the number of tiles. It's row, col.
Assumes that all work is done by the same process.
This orders the work so that it uses cache better.
}
\description{
Takes the extent of the tiles and returns a list of which tiles to do.
}
