% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{task_work_from_tiles}
\alias{task_work_from_tiles}
\title{Takes the extent of the tiles and returns a list of which tiles to do.}
\usage{
task_work_from_tiles(tiles)
}
\arguments{
\item{tiles}{A list of tiles as a 2xN matrix.}
}
\value{
An array that's 2 x the number of tiles. It's row, col.
Assumes that all work is done by the same process.
This orders the work so that it uses cache better.
}
\description{
Takes the extent of the tiles and returns a list of which tiles to do.
}
