% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{worker}
\alias{worker}
\title{Reads the plan and produces this task's part of that plan.}
\usage{
worker(args)
}
\arguments{
\item{args}{Command-line arguments. If you don't give it a task, it
reads that task ID from the \code{SGE_TASK_ID} environment variable.

If you want to call this, see \code{\link{main}} for an example.}
}
\description{
Reads the plan and produces this task's part of that plan.
}
