% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_variables.R
\name{years_in_filenames}
\alias{years_in_filenames}
\title{Given a bunch of filenames, find all four-number digit strings.}
\usage{
years_in_filenames(filenames)
}
\arguments{
\item{filenames}{A list of character strings to search.}
}
\value{
A list of numeric years.
}
\description{
Given a bunch of filenames, find all four-number digit strings.
}
