% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_variables.R
\name{data_for_country}
\alias{data_for_country}
\title{Get all the data for a country as a convenience function.}
\usage{
data_for_country(country_alpha3, years)
}
\description{
Get all the data for a country as a convenience function.
}
