% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc_kappa.R
\name{find_relative_extent}
\alias{find_relative_extent}
\title{Find where the tile is within the process extent.
Both are relative to the domain extent.}
\usage{
find_relative_extent(process_extent, tile_extent)
}
\description{
Find where the tile is within the process extent.
Both are relative to the domain extent.
}
