% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_work.R
\name{pixel_two}
\alias{pixel_two}
\title{This was made with fits to the middle values of PfPR.
It was finished Monday 23 Nov 2020. This was the version sent
to MAP that had reasonable values (c is corrected in last step).}
\usage{
pixel_two(pfpr, am, params, strategies)
}
\description{
This was made with fits to the middle values of PfPR.
It was finished Monday 23 Nov 2020. This was the version sent
to MAP that had reasonable values (c is corrected in last step).
}
