% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extents.R
\name{remove_tiles_over_water}
\alias{remove_tiles_over_water}
\title{Removes tiles that have all NA for this example slice.}
\usage{
remove_tiles_over_water(available)
}
\arguments{
\item{available}{This is a description of where all tiles are.}
}
\value{
A list of remaining tiles, by row and column, so a 2D array.
Tile is within the domain. The sample data is already cut
to that domain.
}
\description{
Removes tiles that have all NA for this example slice.
}
