% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_variables.R
\name{write_output}
\alias{write_output}
\title{Given outputs for a tile, write them to a folder.}
\usage{
write_output(output, years, domain_dimensions, domain_extent, args, options)
}
\arguments{
\item{output}{A list of results.}

\item{years}{Years that will save.}

\item{domain_dimensions}{Resolution of the raster and bounds in lat-long.}

\item{domain_extent}{Bounds and resolution of the raster.}

\item{args}{Command-line arguments that have been parsed.}

\item{options}{Settings from the toml file.}
}
\description{
Given outputs for a tile, write them to a folder.
}
